## Summary

Updated the cost model to use a fixed cost for map operations rather than per-element cost. Also added benchmarks to help compare the cost of these with each other.


<details><summary>apple silicon arm64 benchmarks</summary>
<p>

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Apple M1 Pro
                                                    │ StringEquality │  StringHasPrefix   │   SliceContains    │    MapContains     │
                                                    │     sec/op     │   sec/op     vs base   │   sec/op     vs base   │   sec/op     vs base   │
CostEstimation/flavour=Equal_8chars-10                   2.056n ± 1%
CostEstimation/flavour=Equal_32chars-10                  2.728n ± 0%
CostEstimation/flavour=Equal_64chars-10                  2.565n ± 0%
CostEstimation/flavour=NotEqual_8chars-10                2.222n ± 0%
CostEstimation/flavour=NotEqual_32chars-10               2.064n ± 1%
CostEstimation/flavour=NotEqual_64chars-10               2.252n ± 0%
CostEstimation/flavour=ShortPrefix_8chars_Match-10                     2.158n ± 0%
CostEstimation/flavour=LongPrefix_32chars_Match-10                     2.720n ± 1%
CostEstimation/flavour=NearMiss_LastChar_32times-10                    2.400n ± 1%
CostEstimation/flavour=Size_1-10                                                            2.239n ± 0%
CostEstimation/flavour=Size_2-10                                                            3.175n ± 1%          7.053n ± 2%
CostEstimation/flavour=Size_8-10                                                            5.050n ± 0%
CostEstimation/flavour=Size_16-10                                                           7.557n ± 1%          9.596n ± 6%
CostEstimation/flavour=Size_32-10                                                                                9.556n ± 6%
CostEstimation/flavour=Size_128-10                                                                               9.791n ± 7%
CostEstimation/flavour=Size_256-10                                                                               9.631n ± 8%
geomean                                                  2.301n        2.415n       ? ¹ ²   4.058n       ? ¹ ²   9.058n       ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
```

</p>
</details>